A good progress, here's what I did: - added config.qs.yaml -> This file tells qs where to look for the entry file - modified modules/Bar.qml -> Main bar - added Clock.qml -> This one is the clock, simple enough - added modules/Workspaces.qml -> For workspaces (unfinished)
Note:
Quickshell is hard, very hard. But I'm still very curious about it's full potential and I want to know more. I still have no idea what I'm doing, I thought it would be just as hard as configuring hyprland and waybar for the first time, but it's actually way harder. I don't know if I will continue or not, I'm not sure about learning a whole new language that I'm not familiar with just to customize a desktop.

There's still a bunch of things that I have to fix, here are some of them:
- Workspace
- Bar dynamic layout
- Utility modules such as volume control, brightness, etc.